The initial open source port of CreateMachineState() caused a problem when adding the ACL for Everyone to the registry key.  The SD returned by RegGetKeySecurity() cannot be modified in place, and a copy has to be made first.
The key created by this method has a different location depending on bit width
of Windows and the bit width of the application:

-----------------------------------------------
HKLM\Software\Google\Common\Rlz
HKLM\Software\Wow6432Node\Google\Common\Rlz
-----------------------------------------------

TEST=make sure that an admin user can correctly create the key and change
its security to allow everyone to read and write to the key.  Test this by
logging in with a non-admin account and try adding a value to the key.

Review URL: http://codereview.appspot.com/1958045