Implement RlzValueStoreMac.
It writes to an in-memory dictionary that's read from disk when the
store lock is acquired and that writes to disk the that lock is released.

The lock implementation itself will be done in another CL.

Since all tests do now pass, enable them on mac.

BUG=none
TEST=unit tests

Review URL: https://codereview.appspot.com/5885047