Move GetMachineId() to its own file.
The sha1 logic is now shared cross-platform which calls out to a
platform-specific function to get system data.

GetMachineId() now returns a string instead of a wstring. The SID
is still a string16 instead of a wstring, which is the same on windows.
(I considered making it a string too, but considered it too risky since
if a wide character creeps in somehow it'd change the SHA result.)

The platform-specific implementation on mac is still empty.

BUG=none
TEST=none

Review URL: https://codereview.appspot.com/5877053