mac: Make ScopedRlzValueStoreLock handle nested lock failures.
I thought that lib_rlz always early-exits when it fails to acquire the rlz lock,
and so nested lock acquisition failures should never happen. That is true as
long as SupplementaryBranding isn't used, and that can even be used from
user code like so:

  {
    rlz_lib::SupplementaryBranding branding(reactivation_brand.c_str());
    ret &= rlz_lib::RecordProductEvent(product, point, event_id);
  }

Since SupplementaryBranding tries to acquire the rlz lock and keeps it while
it's on the scope, if RecordProductEvent fails to acquire the lock it'll be a
nested lock failure. Update the code to not CHECK on this case.

Add a test that covers this scenario.

BUG=chromium:121255
TEST=unit test. I verified the test crashes in the CHECK without my code change.

Review URL: https://codereview.appspot.com/5966066